listing the datas
