+++ backend/index.js
@@ -5,7 +5,6 @@ import user from "./models/user.js";
 import Project from "./models/project.js";
 import userRouter from "./router/userRouter.js";
 import projectRouter from "./router/projectRouter.js";
-import benificiaryRouter from "./router/benificiaryrouter.js";
 
 
 dotenv.config();
@@ -31,6 +30,5 @@ app.use(express.json());
 
 app.use("/user", userRouter);
 app.use("/project",projectRouter);
-app.use("/benificiary",benificiaryRouter);
 
 app.listen(process.env.PORT, console.log(`Hello from port: ${process.env.PORT}`));
