--- backend/router/benificiaryrouter.js
@@ -1,9 +1,0 @@
-import { Router } from "express";
-import { listBenificiary } from "../controller/benificiary.js";
-
-
-const benificiaryRouter = Router();
-
-benificiaryRouter.route("/").get(listBenificiary);
-
-export default benificiaryRouter;
