+++ backend/router/userRouter.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { listUser, listUserById, login, register, userDelete, userUpdate } from "../controller/userController.js";
+import { listAidAgency, listBenificiary, listDonor, listUser, listUserById, login, register, userDelete, userUpdate } from "../controller/userController.js";
 import validateData from "../middlewares/validation.js";
 import userSchema from "../schemas/user.js";
 
@@ -9,5 +9,8 @@ userRouter.route("/signup").post(validateData(userSchema), register);
 userRouter.route("/").get(listUser).put(validateData(userSchema), userUpdate);
 userRouter.route("/:id").get(listUserById).delete(userDelete);
 userRouter.route("/login").post(login);
+userRouter.route("/list/aidagency").get(listAidAgency);
+userRouter.route("/list/donor").get(listDonor);
+userRouter.route("/list/benificiary").get(listBenificiary);
 
 export default userRouter;
